replace the old DefaultDecisionMaker by SimpleDecisionMaker, but use "DefaultDecisionMaker" as its class name
